Document more components
