+++ components/portfolio/Contributor.stories.mdx
@@ -1,0 +1,20 @@
+import { Canvas, Meta, Story, ArgsTable } from "@storybook/addon-docs";
+
+import Contributor from "./Contributor";
+
+export const Template = (args) => <Contributor {...args} />;
+
+<Meta title="Design-System/Portfolio/Contributor" component={Contributor} />
+
+# Contributor
+
+Used for projects to show the contributors. Takes a GitHub username under `username` and a name under `name`
+
+<Canvas>
+  <Story
+    name="Default"
+    args={{ username: "samrobbins85", name: "Sam Robbins" }}
+  />
+</Canvas>
+
+<ArgsTable of={Contributor} story="Default" />
