+++ components/portfolio/Contributor.tsx
@@ -1,6 +1,4 @@
-import Image from "next/image";
-
-export default function Coder({
+export default function Contributor({
   username,
   name,
 }: {
@@ -10,13 +8,10 @@ export default function Coder({
   return (
     <div className="flex items-center gap-x-2" key={username}>
       <div className="inline-block h-10 w-10 relative">
-        <Image
+        <img
           className="rounded-full"
           src={`https://github.com/${username}.png`}
           alt={username}
-          layout="fill"
-          objectFit="contain"
-          priority
         />
       </div>
       <div className="flex flex-col">
