+++ components/portfolio/GitHubButton.stories.mdx
@@ -1,0 +1,52 @@
+import { Canvas, Meta, Story, ArgsTable } from "@storybook/addon-docs";
+
+import GitHubButton from "./GitHubButton";
+
+export const Template = (args) => <GitHubButton {...args} />;
+
+<Meta title="Design-System/Portfolio/GitHubButton" component={GitHubButton} />
+
+# GitHubButton
+
+This comes in two variants, single and multiple. Single is for when a project uses only one repository, and so clicking it takes you straight to the repository. For multiple, a dropdown will be presented of repositories that can be chosen
+
+<Canvas>
+  <Story
+    name="Single"
+    decorators={[
+      (Story) => (
+        <div className="flex justify-center">
+          <Story />
+        </div>
+      ),
+    ]}
+    args={{
+      repos: ["https://github.com/samrobbins85/samrobbins.uk"],
+    }}
+  />
+</Canvas>
+
+## Args Table
+
+<ArgsTable of={GitHubButton} story="Single" />
+
+## Multiple
+
+<Canvas>
+  <Story
+    name="Multiple"
+    decorators={[
+      (Story) => (
+        <div className="flex justify-center">
+          <Story />
+        </div>
+      ),
+    ]}
+    args={{
+      repos: [
+        "https://github.com/samrobbins85/samrobbins.uk",
+        "https://github.com/samrobbins85/latex-tailwind",
+      ],
+    }}
+  />
+</Canvas>
