+++ components/writing/Charts/LineChart.stories.mdx
@@ -1,0 +1,53 @@
+import { Canvas, Meta, Story, ArgsTable } from "@storybook/addon-docs";
+
+import LineChart from "./LineChart";
+
+export const Template = (args) => <LineChart {...args} />;
+
+<Meta title="Design-System/Writing/Charts/LineChart" component={LineChart} />
+
+# LineChart
+
+This takes two sets of data `data` and `data1` and shows them in a line chart to compare them
+
+<Canvas>
+  <Story
+    name="Default"
+    decorators={[
+      (Story) => (
+        <div className="mx-auto prose prose-serif dark:prose-light dark:prose-serifLight">
+          <Story />
+        </div>
+      ),
+    ]}
+    args={{
+      data: [
+        { x: 10, y: 158 },
+        { x: 20, y: 287 },
+        { x: 30, y: 438 },
+        { x: 40, y: 572 },
+        { x: 50, y: 708 },
+        { x: 60, y: 856 },
+        { x: 70, y: 1077 },
+        { x: 80, y: 1157 },
+      ],
+      data1: [
+        { x: 10, y: 106 },
+        { x: 20, y: 213 },
+        { x: 30, y: 323 },
+        { x: 40, y: 428 },
+        { x: 50, y: 574 },
+        { x: 60, y: 646 },
+        { x: 70, y: 757 },
+        { x: 80, y: 870 },
+      ],
+      dataName: "JavaScript",
+      data1Name: "Web Assembly",
+      xlabel: "Time (ms)",
+      xlabel: "Resolution (MP)",
+      caption: "Relationship between time and resolution",
+    }}
+  />
+</Canvas>
+
+<ArgsTable of={LineChart} story="Default" />
