+++ components/writing/chart.ts
@@ -1,0 +1,4 @@
+export { default as BarChart } from "./Charts/BarChart";
+export { default as LineChart } from "./Charts/LineChart";
+export { default as BarMulti } from "./Charts/MultiBarChart";
+export { default as PieChart } from "./Charts/PieChart";
